[RENAME] IAM으로 컨텍스트명 변경
