+++ app/iam-shared-kernel/settings.gradle
