+++ app/iam-shared-kernel/src/main/java/com/chinda/user_shared_kernel/model/Platform.java
