+++ app/iam/src/main/java/com/chinda/user/application/exceptions/UserNotRegisteredException.java
