+++ app/iam/src/main/java/com/chinda/user/domain/access/dto/OAuthTokenResponse.java
@@ -1,0 +1,5 @@
+package com.chinda.user.domain.access.dto;
+
+public interface OAuthTokenResponse {
+    public String getAccessToken();
+}
