--- app/user/src/main/java/com/chinda/user/domain/iam/OAuthRequester.java
@@ -1,6 +1,0 @@
-package com.chinda.user.domain.iam;
-
-public interface OAuthRequester {
-    OAuthAgreedUser getOAuthAgreedUser(String authCode);
-
-}
