--- app/user/src/main/java/com/chinda/user/domain/iam/dto/OAuthTokenResponse.java
@@ -1,5 +1,0 @@
-package com.chinda.user.domain.iam.dto;
-
-public interface OAuthTokenResponse {
-    public String getAccessToken();
-}
