--- app/user/src/main/java/com/chinda/user/domain/iam/dto/OAuthUserResponse.java
@@ -1,5 +1,0 @@
-package com.chinda.user.domain.iam.dto;
-
-public interface OAuthUserResponse {
-    public Long getSocialId();
-}
